requirements.txt
django==4.2
djangorestframework
Pillow
python-telegram-bot
